# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stadion Tracker API",
        "description": "API-documentatie voor stadionbezoeken, wedstrijden, posts en gebruikersinteractie.",
        "contact": {
            "email": "cedricvanhoorebeke@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.ddev.site",
            "description": "DDEV ontwikkelomgeving"
        }
    ],
    "paths": {
        "/follow": {
            "post": {
                "tags": [
                    "Follows"
                ],
                "summary": "Volg een team, stadion of gebruiker",
                "operationId": "c445351cf4039ea6ab8ee80d592f231a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "followable_type",
                                    "followable_id"
                                ],
                                "properties": {
                                    "followable_type": {
                                        "type": "string",
                                        "enum": [
                                            "team",
                                            "stadium",
                                            "user"
                                        ],
                                        "example": "team"
                                    },
                                    "followable_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesvol gevolgd",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Succesvol gevolgd."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Je volgt dit al."
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/unfollow": {
            "delete": {
                "tags": [
                    "Follows"
                ],
                "summary": "Ontvolg een team, stadion of gebruiker",
                "operationId": "eba2e71b1d57e6e911418f640bfad4a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "followable_type",
                                    "followable_id"
                                ],
                                "properties": {
                                    "followable_type": {
                                        "type": "string",
                                        "enum": [
                                            "team",
                                            "stadium",
                                            "user"
                                        ],
                                        "example": "stadium"
                                    },
                                    "followable_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ontvolgd",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ontvolgd."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/following": {
            "get": {
                "tags": [
                    "Follows"
                ],
                "summary": "Lijst van gevolgde items door de gebruiker",
                "operationId": "b74642416479ded7f45ca1d0210cc7dd",
                "responses": {
                    "200": {
                        "description": "Lijst met volgitems",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "follows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "followable_type": {
                                                        "type": "string"
                                                    },
                                                    "followable_id": {
                                                        "type": "integer"
                                                    },
                                                    "followable": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Toon alle gebruikers",
                "operationId": "7208c29037df2b8bb02499ba39ed8a78",
                "responses": {
                    "200": {
                        "description": "Lijst van gebruikers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Toon één gebruiker op ID",
                "operationId": "6946bbab7898f59a09e3c8e529a38907",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID van de gebruiker",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details van een gebruiker",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gebruiker niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Toon huidige ingelogde gebruiker",
                "operationId": "de3936ad40be9d99a380e4fcfcc59871",
                "responses": {
                    "200": {
                        "description": "Eigen gebruikersprofiel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niet ingelogd"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/profile": {
            "post": {
                "tags": [
                    "Profiel"
                ],
                "summary": "Update het profiel van de ingelogde gebruiker",
                "operationId": "34ed39829ad0eb744b2283606c6fccef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "newusername"
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "banner_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profiel succesvol bijgewerkt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "username",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cédric Van Hoorebeke"
                    },
                    "username": {
                        "type": "string",
                        "example": "cedrictest"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "cedric@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "https://s3-url.com/image.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "example": "https://s3-url.com/banner.jpg",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}